requirements.txt
streamlit>=1.28
requests>=2.25
Pillow>=9.0
openai>=1.3.0